refactor: Centralize queries in a single file
